+++ packages/ibm-products/src/components/index.js
@@ -87,5 +87,14 @@ export { InlineEdit } from './InlineEdit';
 export { EditInPlace } from './EditInPlace';
 export { Guidebanner } from './Guidebanner';
 export { NonLinearReading } from './NonLinearReading';
+
+export { Coachmark } from './Coachmark';
+export { CoachmarkFixed } from './CoachmarkFixed';
+export { CoachmarkBeacon } from './CoachmarkBeacon';
+export { CoachmarkButton } from './CoachmarkButton';
+export { CoachmarkOverlayElements } from './CoachmarkOverlayElements';
+export { CoachmarkOverlayElement } from './CoachmarkOverlayElement';
+export { CoachmarkStack } from './CoachmarkStack';
+
 export { InlineTip } from './InlineTip';
 export { Checklist } from './Checklist';