+++ packages/security/src/components/Header/HeaderListItem/HeaderListItem.js
@@ -10,7 +10,7 @@ import { defaultProps, namespace, propTypes } from './constants';
 
 /**
  * Header list item component.
- * @param {object.<string, *>} props Header list item props.
+ * @param {Record<string, any>} props Header list item props.
  * @returns {HeaderListItem} Header list item instance.
  */
 const HeaderListItem = ({ children, className }) => (