+++ packages/security/src/components/Shell/Shell.js
@@ -129,23 +129,23 @@ Shell.defaultProps = {
 };
 
 Shell.propTypes = {
-  /** @type {object<object.Object>} An object list of header properties. */
+  /** @type {Record<object, object>} An object list of header properties. */
   header: PropTypes.shape(headerPropTypes),
 
-  /** @type {object<object.Object>} An object list of profile properties. */
+  /** @type {Record<object, object>} An object list of profile properties. */
   profile: headerPropTypes.profile,
 
   /** @type {Array<{id: string, tooltip: string, render: Function: React.Element}>} An object list to render custom addon icons. */
   renderAddons: Toolbar.propTypes.renderAddons,
 
-  /** @type {object<object.Object>} An object list of banner properties. */
+  /** @type {Record<object, object>} An object list of banner properties. */
   returnToBanner: PropTypes.shape({
     application: PropTypes.string.isRequired,
     view: PropTypes.string,
     href: PropTypes.string.isRequired,
   }),
 
-  /** @type {object.<string, string>} Object of 'skip-to-content' link information. */
+  /** @type {Record<string, string>} Object of 'skip-to-content' link information. */
   skipToContent: PropTypes.shape({
     /** @type {string} Label text for 'skip-to-content'. */
     label: PropTypes.string,
@@ -154,7 +154,7 @@ Shell.propTypes = {
     href: PropTypes.string,
   }),
 
-  /** @type {object<object.Object>} An object list of toolbar properties. */
+  /** @type {Record<object, object>} An object list of toolbar properties. */
   toolbar: PropTypes.shape(Toolbar.propTypes),
 };